restructure messaging
