--- apps/client/src/pages/room/[roomCode]/components/Room/components/ChatSection/components/index.ts
@@ -1,4 +1,0 @@
-import Message from './Message';
-import ChatHeader from './ChatHeader';
-
-export { Message, ChatHeader };
