+++ apps/client/src/pages/room/[roomCode]/components/Room/components/Sidebar/Sidebar.module.css
@@ -1,0 +1,14 @@
+.container {
+	--header-height: 4.5rem;
+	--message-form-height: 5rem;
+	--messages-list-height: calc(100% - var(--header-height) - var(--message-form-height));
+	--base-padding: 1rem;
+
+	/* extra width so primary-panel-bg is visible a bit to the the right.
+	Needed for when cubic-bezier transition moves panel to left a bit on close */
+	width: calc(100% + 2rem);
+	padding-right: 2rem;
+	background-color: var(--primary-panel-bg);
+	height: 100%;
+}
+
