+++ apps/client/src/pages/room/[roomCode]/components/Room/components/Sidebar/components/MessagesList/MessagesList.module.css
@@ -1,0 +1,5 @@
+.messagesList {
+	height: var(--messages-list-height);
+	overflow-y: auto;
+	padding-inline: 0.5rem;
+}
