+++ apps/client/src/pages/room/[roomCode]/components/Room/components/Sidebar/components/index.ts
@@ -1,0 +1,5 @@
+import MessagesList from './MessagesList';
+import ChatHeader from './ChatHeader';
+import MessageInput from './MessageInput';
+
+export { MessagesList, ChatHeader, MessageInput };
